+++ frontend/src/metabase/components/LoadingAndErrorWrapper/LoadingAndErrorWrapper.tsx
@@ -127,6 +127,7 @@ export const LoadingAndErrorWrapper = forwardRef<
     }
     return Children.only(children);
   }
+
   return (
     <div className={className} style={style} data-testid={testId} ref={ref}>
       {error ? (
